workflows/disease_prediction/requirements.txt
Pillow~=9.5.0
PyYAML~=6.0
intel-tensorflow==2.12.0
neural-compressor==2.1.1
numpy~=1.23.5
tensorflow-hub~=0.13.0
